requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
platformdirs==4.4.0
Pygments==2.19.2
python-mpd2==3.1.1
qobuz==0.0.2
requests==2.32.5
rich==14.1.0
textual==6.1.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.5.0
